Do nothing if coverage file not found

+++ test/patch_lcov.py
@@ -3,16 +3,17 @@ import re
 from pathlib import Path
 
 path = Path("coverage.lcov")
-lines = path.read_text().splitlines()
-modified_lines = []
-for line in lines:
-    if line.startswith("SF:"):
-        m = re.search(r"^(SF:).*(sqlfluff/.*)", line)
-        if m:
-            modified_lines.append(f"{m.group(1)}src/{m.group(2)}")
+if path.exists():
+    lines = path.read_text().splitlines()
+    modified_lines = []
+    for line in lines:
+        if line.startswith("SF:"):
+            m = re.search(r"^(SF:).*(sqlfluff/.*)", line)
+            if m:
+                modified_lines.append(f"{m.group(1)}src/{m.group(2)}")
+            else:
+                print(f"Could not patch line: {line}")
+                modified_lines.append(line)
         else:
-            print(f"Could not patch line: {line}")
             modified_lines.append(line)
-    else:
-        modified_lines.append(line)
-path.write_text("\n".join(modified_lines))
+    path.write_text("\n".join(modified_lines))
